+++ items.lua
@@ -3,4 +3,8 @@ PlaceObj('ModItemCode', {
 	'name', "CAI_1Init",
 	'FileName', "Code/CAI_1Init.lua",
 }),
+PlaceObj('ModItemCode', {
+	'name', "CAI_2ModConfig",
+	'FileName', "Code/CAI_2ModConfig.lua",
+}),
 }